requirements.txt
arrow==0.12.0
backports.functools-lru-cache==1.2.1
certifi==2017.11.5
chardet==3.0.4
elasticsearch==6.0.0
idna==2.6
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
schedule==0.5.0
six==1.11.0
urllib3==1.22
## The following requirements were added by pip freeze:
